+++ dvr_scan/__init__.py
@@ -49,7 +49,7 @@ from dvr_scan.scanner import VideoLoadFailure
 
 
 # Used for module identification and when printing copyright & version info.
-__version__ = 'v1.4'
+__version__ = 'v1.5-dev'
 
 # About & copyright message string shown for the -v/--version CLI argument.
 ABOUT_STRING = """-----------------------------------------------