+++ dvr_scan/scanner.py
@@ -509,6 +509,7 @@ class ScanContext(object):
 
         # Length of buffer we require in memory to keep track of all frames required for -l and -tb.
         buff_len = pre_event_len + min_event_len
+        frames_since_last_event = 0
 
         # Motion event scanning/detection loop. Need to avoid CLI output/logging until end of the
         # main scanning loop below, otherwise it will interrupt the progress bar.
@@ -577,6 +578,7 @@ class ScanContext(object):
                     num_frames_post_event += 1
                     if num_frames_post_event >= post_event_len:
                         in_motion_event = False
+                        frames_since_last_event = 0
                         # The event ended on the previous frame, so correct for that.
                         event_end = presentation_time
                         # The duration, however, should include the PTS of the end frame.
@@ -587,6 +589,7 @@ class ScanContext(object):
                             self._video_writer.release()
                             self._video_writer = None
             else:
+                frames_since_last_event += 1
                 if not self._scan_only:
                     self._draw_overlays(frame_rgb_origin, presentation_time)
                     buffered_frames.append(frame_rgb_origin)
@@ -596,9 +599,10 @@ class ScanContext(object):
                     in_motion_event = True
                     event_window = []
                     num_frames_post_event = 0
-                    shifted_start = max(0, self._curr_pos.frame_num - (pre_event_len + min_event_len))
-                    if shifted_start < 0:
-                        shifted_start = 0
+                    shift_amount = min(frames_since_last_event, pre_event_len + min_event_len)
+                    if self._frame_skip:
+                        shift_amount *= (self._frame_skip + 1)
+                    shifted_start = max(0, self._curr_pos.frame_num - shift_amount)
                     event_start = FrameTimecode(shifted_start, self._video_fps)
                     # Open new VideoWriter if needed, write buffered_frames to file.
                     if not self._scan_only: